requirements.txt
transformers==4.52.4
bitsandbytes==0.46.0
accelerate==1.7.0
langchain==0.3.25
langchainhub==0.1.21
pymupdf==1.25.0
faiss-cpu==1.8.0
langchain_experimental==0.3.4
langchain-community==0.3.24
langchain-huggingface==0.2.0
python-dotenv==1.1.0
pypdf
streamlit==1.36.0
torch
bitsandbytes
accelerate
langchain_google_genai